Introduce params for accepting ACLs
Added the following params for cluster registration
owner_info: User can update the owner_info through this
access_control_list: A list where ACLs for users/teams can be provided

The provider fetches the users and teams information through an API call.
It validates the provided user and team information with the information
fetched from the cluster.
The provider framework does the other validation of type and permission and
optional parameters.
Post a successful validation, the accessControlList object is added to the data
in the POST request for registering the cluster